testproj/requirements.txt
dj-rest-auth==2.1.3
Django==3.1.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
